change the configuration structure
