+++ GANDLF/Configuration/parameters.py
@@ -1,11 +1,10 @@
 from pydantic import BaseModel, ConfigDict
-from GANDLF.Configuration.default_parameters import DefaultParameters
-from GANDLF.Configuration.user_defined_parameters import UserDefinedParameters
+from GANDLF.Configuration.Parameters.user_defined_parameters import UserDefinedParameters
 
 
 class ParametersConfiguration(BaseModel):
     model_config = ConfigDict(extra="allow")
 
 
-class Parameters(ParametersConfiguration,UserDefinedParameters):
+class Parameters(ParametersConfiguration, UserDefinedParameters):
     pass
